Remove Brakeman gem since Rails 8+ includes it by default
